chore(packages/testcontainer): Refactor testcontainers code structure (#1906)
<!--  Thanks for sending a pull request! -->

#### What this PR does / why we need it:

Refactor code in https://github.com/JellyfishSDK/jellyfish/pull/1790 to
simplify `NativeChainContainer.ts`

#### Additional comments?:

Fixed change requests:
-
https://github.com/JellyfishSDK/jellyfish/pull/1790#discussion_r1038748955
-
https://github.com/JellyfishSDK/jellyfish/pull/1790#discussion_r1038749032
-
https://github.com/JellyfishSDK/jellyfish/pull/1790#discussion_r1039171512
