+++ packages/testcontainers/__tests__/containers/NativeChainContainer/Masternode.test.ts
@@ -18,27 +18,27 @@ describe('nativechain masternode', () => {
   })
 
   it('should generate and wait for block', async () => {
-    await container.generate(4)
-    const count = await container.getBlockCount()
+    await container.rpc.generate(4)
+    const count = await container.rpc.getBlockCount()
     expect(count).toBeGreaterThan(3)
   })
 
   it('should waitForGenerate', async () => {
-    await container.waitForGenerate(100)
-    const count = await container.getBlockCount()
+    await container.waitFor.generate(100)
+    const count = await container.rpc.getBlockCount()
     expect(count).toBeGreaterThanOrEqual(100)
   })
 
   it('should getBestBlockHash', async () => {
-    await container.waitForGenerate(100)
-    const hash = await container.getBestBlockHash()
+    await container.waitFor.generate(100)
+    const hash = await container.rpc.getBestBlockHash()
     expect(hash.length).toStrictEqual(64)
   })
 
   it('should waitForBlockHeight', async () => {
-    const bc1 = await container.getBlockCount()
-    await container.waitForBlockHeight(bc1 + 100)
-    const bc2 = await container.getBlockCount()
+    const bc1: number = await container.rpc.getBlockCount()
+    await container.waitFor.blockHeight(bc1 + 100)
+    const bc2 = await container.rpc.getBlockCount()
     expect(bc2).toStrictEqual(bc1 + 101)
   })
 })
